refactor(cli): split bin lib utilities and centralize shared loaders
