+++ packages/cli/src/bin/version.ts
@@ -1,0 +1,1 @@
+export const CLI_VERSION = '0.1.0'
