+++ packages/core/src/index.ts
@@ -6,5 +6,7 @@ export {
 } from './canonical.js'
 export { planDiff } from './planner.js'
 export { createSnapshot } from './snapshot.js'
+export { loadSchemaDefinitions, type SchemaLoaderOptions } from './schema-loader.js'
+export { normalizeSQLFragment } from './sql-normalizer.js'
 export { toCreateSQL } from './sql.js'
 export { assertValidDefinitions, validateDefinitions } from './validate.js'
