+++ packages/core/src/sql-normalizer.ts
@@ -1,0 +1,3 @@
+export function normalizeSQLFragment(value: string): string {
+  return value.replace(/\s+/g, ' ').trim()
+}
